+++ src/main.py
@@ -1,10 +1,7 @@
-import os
-from fastapi import FastAPI, Query
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-import sqlite3
-from typing import List, Dict
+from src.routes import generos_routes, personas_routes, preferencias_routes
 
-from src.models.preferencia import PreferenciaDTO
 
 '''
 Correr con:
@@ -21,68 +18,9 @@ app.add_middleware(
     allow_headers=["*"],  # permitir todos los headers
 )
 
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))  
-DB_PATH = os.path.join(BASE_DIR, "data", "recomendador.sqlite")
-
-#GENEROS
-
-@app.get("/generos")
-def obtener_generos():
-    conn = sqlite3.connect(DB_PATH)
-    cursor = conn.cursor()
-    cursor.execute("SELECT id, nombre FROM genero")
-    generos = [{"id": row[0], "nombre": row[1]} for row in cursor.fetchall()]
-    conn.close()
-    return generos
-
-@app.get("/actores")
-def get_actores(busqueda: str = Query("", description="Texto para filtrar actores por nombre")):
-    """
-    Devuelve todas las personas que tengan la profesión 'actor' (id_profesion=1),
-    opcionalmente filtradas por nombre.
-    """
-    return buscar_personas_por_profesion("1", busqueda)
+app.include_router(generos_routes.router)
+app.include_router(personas_routes.router)
+app.include_router(preferencias_routes.router)
 
 
-@app.get("/directores")
-def get_directores(busqueda: str = Query("", description="Texto para filtrar directores por nombre")):
-    """
-    Devuelve todas las personas que tengan la profesión 'director' (id_profesion=3),
-    opcionalmente filtradas por nombre.
-    """
-    return buscar_personas_por_profesion("3", busqueda)
 
-#RECIBIR PREFERENCIA DEL FRONTEND
-@app.post("/preferencias")
-def recibir_preferencias(preferencia: PreferenciaDTO):
-    print("📩 Preferencias recibidas:")
-    print(preferencia.model_dump())  # Versión en dict
-    return {"mensaje": "Preferencias recibidas correctamente"}
-
-def buscar_personas_por_profesion(id_profesion: str, busqueda: str = "") -> List[Dict]:
-    """
-    Busca personas que tengan la profesión indicada (id_profesion)
-    y opcionalmente filtradas por nombre.
-    """
-    conn = sqlite3.connect(DB_PATH)
-    conn.row_factory = sqlite3.Row
-    cursor = conn.cursor()
-
-    query = """
-        SELECT DISTINCT p.id, p.nombre
-        FROM persona p
-        JOIN profesion_titulo pt ON p.id = pt.id_persona
-        WHERE pt.id_profesion = ?
-    """
-
-    params = [id_profesion]
-
-    if busqueda:
-        query += " AND p.nombre LIKE ?"
-        params.append(f"%{busqueda}%")
-
-    cursor.execute(query, params)
-    resultados = [dict(row) for row in cursor.fetchall()]
-
-    conn.close()
-    return resultados
